influx/requirements.txt
confluent_kafka==2.5.0
influxdb_client==1.45.0
python-dotenv==1.0.1